+++ CHANGELOG.md
@@ -5,6 +5,8 @@ the [Keep a CHANGELOG](http://keepachangelog.com/)
 principles.
 
 ## Unreleased
+- Fixed: Import issue where text import of non-string values, like numbers would break squanto translations.
+- Fixed: Import issue where menu headers are not labeled as expected. Now throws an error on import.
 
 ## 0.8.20 - 2024-07-22
 -   Fixed: Could not save slug containing slashes on a nestable page.